float to uint

+++ lab3/Core/Src/main.c
@@ -52,7 +52,7 @@ float BeforeGearRatio;
 float MotorReadRPM;
 float microsectominute;
 
-float MotorSetDuty = 50;
+uint16_t MotorSetDuty = 50;
 float CompareValue = 0;
 float PreviousCompareValue = 0;
 
@@ -454,8 +454,8 @@ uint8_t CheckMotorControlEnable(){
 
 	case(0):
 		  // (1) PWM Duty Cycle
-		  // Compare Value = Duty cycle of PWM * Time Period
-		  CompareValue = MotorSetDuty*10; // counter period 1000 but input 0 - 100
+		  // counter period 1000 but input 0 - 100
+		  CompareValue = MotorSetDuty*10;
 		  __HAL_TIM_SET_COMPARE(&htim1, TIM_CHANNEL_1,CompareValue); // change reference output compare
 		  break;
 	}
